requirements.txt
PySide6-Fluent-Widgets
nuitka